Inject CardReaderManager
