+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/cardreader/CardReaderPaymentViewModel.kt
@@ -48,6 +48,7 @@ private const val ARTIFICIAL_RETRY_DELAY = 500L
 class CardReaderPaymentViewModel @AssistedInject constructor(
     @Assisted savedState: SavedStateWithArgs,
     dispatchers: CoroutineDispatchers,
+    cardReaderManager: CardReaderManager?,
     private val logger: AppLogWrapper,
     private val orderStore: WCOrderStore
 ) : DaggerScopedViewModel(savedState, dispatchers) {
@@ -57,12 +58,13 @@ class CardReaderPaymentViewModel @AssistedInject constructor(
     private val viewState = MutableLiveData<ViewState>(LoadingDataState)
     val viewStateData: LiveData<ViewState> = viewState
 
-    private lateinit var cardReaderManager: CardReaderManager
+    // TODO remove this, and make the constructor parameter as a non nullable property when
+    //  the actual implementation is injected in release builds
+    private val cardReaderManager = cardReaderManager!!
 
     private var paymentFlowJob: Job? = null
 
-    final fun start(cardReaderManager: CardReaderManager) {
-        this.cardReaderManager = cardReaderManager
+    final fun start() {
         // TODO cardreader Check if the payment was already processed and cancel this flow
         // TODO cardreader Make sure a reader is connected
         if (paymentFlowJob == null) {
