+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/CardReaderSettingsFragment.kt
@@ -27,6 +27,7 @@ import com.woocommerce.android.cardreader.CardReaderStatus.NOT_CONNECTED
 import com.woocommerce.android.cardreader.SoftwareUpdateStatus.Installing
 import com.woocommerce.android.databinding.FragmentSettingsCardReaderBinding
 import com.woocommerce.android.extensions.navigateSafely
+import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -35,16 +36,20 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.wordpress.android.util.AppLog
+import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
 
 // TODO cardreader update this comment
 /**
  * This fragment currently contains a UI for testing purposes. It'll be removed before the release.
  */
+@AndroidEntryPoint
 class CardReaderSettingsFragment : Fragment(R.layout.fragment_settings_card_reader), CoroutineScope {
     companion object {
         const val TAG = "card-reader-settings"
     }
+
+    @set:Inject var cardReaderManager: CardReaderManager? = null
 
     protected var job: Job = Job()
 
@@ -88,7 +93,7 @@ class CardReaderSettingsFragment : Fragment(R.layout.fragment_settings_card_read
         binding.updateReaderSoftware.setOnClickListener {
             launch(Dispatchers.Default) {
                 try {
-                    getCardReaderManager()?.let {
+                    cardReaderManager?.let {
                         updateReaderSoftware(it, binding.softwareUpdateStatus)
                     }
                 } catch (e: CancellationException) {
@@ -124,7 +129,7 @@ class CardReaderSettingsFragment : Fragment(R.layout.fragment_settings_card_read
 
     // TODO cardreader move this into a VM
     private fun connectToReader(simulated: Boolean) {
-        getCardReaderManager()?.let { cardReaderManager ->
+        cardReaderManager?.let { cardReaderManager ->
             if (!cardReaderManager.isInitialized) {
                 cardReaderManager.initialize(requireActivity().application)
             }
@@ -144,7 +149,7 @@ class CardReaderSettingsFragment : Fragment(R.layout.fragment_settings_card_read
                         }
                         is ReadersFound -> {
                             if (event.list.isNotEmpty()) {
-                                val success = getCardReaderManager()?.connectToReader(event.list[0]) ?: false
+                                val success = cardReaderManager.connectToReader(event.list[0]) ?: false
                                 Snackbar.make(
                                     requireView(),
                                     "Connecting to reader ${if (success) "succeeded" else "failed"}",
@@ -179,7 +184,4 @@ class CardReaderSettingsFragment : Fragment(R.layout.fragment_settings_card_read
         super.onDestroy()
         job.cancel()
     }
-
-    private fun getCardReaderManager(): CardReaderManager? =
-        (requireActivity().application as? WooCommerce)?.cardReaderManager
 }
